turbo-codes/docker_requirements.txt
scikit-commpy==0.7.0
tqdm==4.62.3
plotly==5.3.1
pandas==1.1.5
kaleido==0.2.1
pytest==6.2.5
git+https://www.github.com/keras-team/keras-contrib.git
tensor_annotations
tensor_annotations_tensorflow_stubs
tensorflow-probability